Separate the algorithm from the data it scans
The raptor folder held both the scan and the structures it scans, so Timetable
sat next to RaptorAlgorithm as though it were part of it rather than the input
to it. The timetable, the calendar and the network they make up move to a
folder of their own, leaving raptor with the algorithm.

Normalise standardises a feed, so it belongs with the rest of the feed handling.
It was reaching backwards into the raptor layer for Interchange and
TransfersByOrigin, which are Record<StopID, ...> feed shapes that had ended up
in RaptorAlgorithm.ts, so those move to GTFS.ts and the dependency goes away.

The connection type was declared in ScanResults and read in results/Connections,
which is the same split that had stopAt living in the module that builds the
network. A connection and how to read it are now one file in raptor, along with
the transfer discriminator that had ended up in ResultsFactory.
