+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./gtfs/GTFS";
+export * from "./gtfs/Normalise";
 export * from "./gtfs/Service";
 export * from "./gtfs/GTFSLoader";
 export * from "./gtfs/TimeParser";
@@ -11,12 +12,12 @@ export * from "./raptor/Queue";
 export * from "./raptor/RaptorAlgorithm";
 export * from "./raptor/RouteScanner";
 export * from "./raptor/ScanResults";
-export * from "./raptor/TripCalendar";
-export * from "./raptor/Network";
-export * from "./raptor/Timetable";
+export * from "./network/TripCalendar";
+export * from "./network/Network";
+export * from "./network/Timetable";
 export * from "./results/Journey";
 export * from "./results/JourneyFactory";
-export * from "./results/Connections";
+export * from "./raptor/Connection";
 export * from "./results/ResultsFactory";
 export * from "./results/filter/MultipleCriteriaFilter";
 export * from "./results/filter/JourneyFilter";
@@ -24,4 +25,3 @@ export * from "./transfer-pattern/results/GraphResults";
 export * from "./transfer-pattern/results/StringResults";
 export * from "./transfer-pattern/TransferPatternRepository";
 export * from "./transfer-pattern/results/TransferPatternResults";
-export { Connection } from "./raptor/ScanResults";
