+++ src/network/TripCalendar.ts
