+++ src/transfer-pattern/results/GraphResults.ts
@@ -1,8 +1,7 @@
-import type { ConnectionIndex } from "../../raptor/ScanResults";
-import type { Network } from "../../raptor/Network";
-import { originIndexOf } from "../../results/Connections";
-import type { StopIdx } from "../../raptor/Timetable";
-import { isTransfer } from "../../results/ResultsFactory";
+import { isTransfer, type ConnectionIndex } from "../../raptor/Connection";
+import type { Network } from "../../network/Network";
+import { originIndexOf } from "../../raptor/Connection";
+import type { StopIdx } from "../../network/Timetable";
 import type { StopID } from "../../gtfs/GTFS";
 import type { Path, TransferPatternResults } from "./TransferPatternResults";
